requirements2.txt
mysql-connector-python==8.1.0
bcrypt==4.0.1

Jinja2==3.1.6
dnspython==2.7.0
docutils==0.21.2
Flask==3.1.0

ipwhois==1.3.0

libpcap==1.11.0b21

matplotlib==3.10.3
numpy==2.2.5

pandas==2.2.3

pyshark==0.6
python-dateutil==2.9.0.post0
python-nmap==0.7.1
python-whois==0.9.5

requests==2.32.3
scapy==2.6.1

WTForms==3.2.1
Flask-Mail==0.10.0
Flask-WTF==1.2.2
email-validator==2.2.0
python-dotenv==1.1.0

SQLAlchemy==2.0.40
alembic==1.13.1
Werkzeug==3.1.3